+++ screens/HomeScreen.js
@@ -2,21 +2,19 @@ import React, {useState, useContext, useEffect} from 'react';
 import { NavigationContext } from 'react-navigation';
 import { NavigationEvents } from "react-navigation";
 import * as SQLite from 'expo-sqlite';
+import QRCodeListItem from '../components/QRCodesListItem.js'
 
 import {
   FlatList,
   YellowBox,
   StyleSheet,
-  Image,
   Text,
-  TouchableOpacity,
-  View,
+  Platform,
 } from 'react-native';
 import { SafeAreaView } from 'react-navigation';
 
 YellowBox.ignoreWarnings(['Failed child context type']);
 
-const trashIcon = require("../assets/images/trash-icon.png")
 const qrDB = SQLite.openDatabase('qrDB');
 
 export default function HomeScreen() {
@@ -50,7 +48,7 @@ export default function HomeScreen() {
 
    const _renderItem = (item) => {
     return (
-      <CodeListItem
+      <QRCodeListItem
         _fetchQRCodes={_fetchQRCodes()}
         item={item.item}
         navigation={navigation}
@@ -76,40 +74,6 @@ export default function HomeScreen() {
   );
 }
 
-class CodeListItem extends React.PureComponent { 
-
-  render() {
-    const openWebview = () => {
-      this.props.navigation.navigate('WebView', {
-         url: this.props.item.url
-      })
-    }
-
-    const deleteItem = () => {
-      qrDB.transaction(
-        tx => {
-          tx.executeSql(`delete from items where id = ?;`, [this.props.item.id]);
-        },
-        null,
-        this.props._fetchQRCodes
-      )
-    }
-
-   return (
-    <TouchableOpacity onPress={() => openWebview()} style={styles.itemContainter}>
-      <View style={styles.idContainer}>
-   <Text style={styles.idText}>{this.props.item.id}</Text>
-      </View>
-      <View style={styles.itemURLContainer}>
-   <Text style={styles.URLText}>{this.props.item.url}</Text>
-      </View>
-      <TouchableOpacity onPress={() => deleteItem()} style={styles.itemDeleteContainer}>
-        <Image source={trashIcon} style={styles.deleteImage}></Image>
-      </TouchableOpacity>
-    </TouchableOpacity>
-    )}
-}
-
 HomeScreen.navigationOptions = {
   header: null,
 };
@@ -118,6 +82,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
+    marginTop: Platform.OS === 'android' ? 20 : 0
   },
   emptyDBText: {
     textAlign: 'center',
@@ -125,38 +90,5 @@ const styles = StyleSheet.create({
     marginLeft: 20,
     marginRight: 20,
     fontSize: 25
-  },
-  itemContainter: {
-    flex: 1, 
-    flexDirection: 'row',
-    backgroundColor: '#D4D4D4',
-    height: 75,
-    marginTop: 3
-  },
-  idContainer: {
-    flex: 0.15,
-    justifyContent: 'center',
-    alignContent: 'center'
-  },
-  idText: {
-    textAlign: 'center',
-    fontWeight: '500',
-    fontSize: 22,
-  },
-  itemURLContainer: {
-    flex: 0.7,
-    justifyContent: 'center'
-  },
-  URLText: {
-    fontSize: 15,
-  },
-  itemDeleteContainer: {
-    flex: 0.15,
-    justifyContent: 'center',
-    alignContent: 'center'
-  },
-  deleteImage: {
-    width: 40,
-    height: 40
-  },
+  }
 });
